requirements.txt
PyQt5>=5.15.0
pyobjc>=9.0
pyobjc-framework-Cocoa>=9.0
